Fixed Temperature::toString precision

+++ src/main/java/app/appmeteo/model/weather/Temperature.java
@@ -16,6 +16,6 @@ public class Temperature {
 
     @Override
     public String toString() {
-        return toCelcius() + "°C";
+        return (int)toCelcius() + "°C";
     }
 }
